+++ cmd/lmm/root.go
@@ -13,7 +13,7 @@ import (
 )
 
 var (
-	version = "0.6.2"
+	version = "0.6.3"
 
 	// Global flags
 	configDir string